requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
click==6.6
decorator==4.0.10
enum34==1.1.6
Flask==0.11.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.12.0
selenium==3.0.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.11
